Split garbage.rs into garbage/ subdirectory (971 -> ~350 lines each)
Split the monolithic garbage.rs (971 lines) into three focused files:
- types.rs (274 lines): Constants, GarbageGrid, WasteProducer, WasteSystem, WasteCollectionGrid
- systems.rs (343 lines): All ECS systems and GarbagePlugin
- tests.rs (365 lines): All unit tests

All pub items re-exported from mod.rs so external imports remain unchanged.

Part of #1389
